Cancel settings hider if settings are changing

+++ SmartPlane/src/main/java/com/tobyrich/app/SmartPlane/PlaneState.java
@@ -43,7 +43,7 @@ public class PlaneState extends Application{
     private int flAssistMode;
     private boolean screenLocked = false;
 
-    public float rudderTrim = 0;
+    public int rudderTrim = 0;
 
     public void setScaler(double scaler) {
         this.scaler = scaler;
